requirements.txt
dash==2.13.0
fake_headers==1.0.2
pandas==2.1.3
python-dotenv==1.0.0
python_dateutil==2.8.2
pytz==2023.3.post1
selenium==4.15.2
SQLAlchemy==2.0.23
psycopg2
dash-bootstrap-components
